refatoracao dos testes, buscando melhorar a estrutura de organizacao
